requirements.txt
requests>=2.32
feedparser>=6.0
beautifulsoup4>=4.12
lxml>=5.0
jinja2>=3.1
python-dotenv>=1.0
